requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pydantic
python-dotenv
passlib
python-jose
requests
pytest
python-multipart
alembic
argon2-cffi
python-whois
scipy
python-magic-bin
celery
redis
